Extract classes; exceptions into packages etc.
