+++ src/main/java/Exceptions/DukeException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public abstract class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+}
