+++ src/main/java/Exceptions/DuplicateTaskNameException.java
@@ -1,0 +1,8 @@
+package Exceptions;
+
+public class DuplicateTaskNameException extends IncorrectInputException {
+
+    public DuplicateTaskNameException(String message) {
+        super(message);
+    }
+}
