+++ src/main/java/Exceptions/IncorrectIndexException.java
@@ -1,0 +1,6 @@
+package Exceptions;
+public class IncorrectIndexException extends IncorrectInputException {
+    public IncorrectIndexException(String message) {
+        super(message);
+    }
+}
