+++ src/main/java/Exceptions/IncorrectInputException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class IncorrectInputException extends DukeException {
+    public IncorrectInputException(String message) {
+        super(message);
+    }
+}
