+++ src/main/java/Exceptions/NotEnoughDatesException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class NotEnoughDatesException extends NotEnoughInputsException {
+    public NotEnoughDatesException(String message) {
+        super(message);
+    }
+}
