+++ src/main/java/Exceptions/NotEnoughInputsException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class NotEnoughInputsException extends IncorrectInputException {
+    public NotEnoughInputsException(String message) {
+        super(message);
+    }
+}
